+++ CHANGELOG.md
@@ -21,3 +21,9 @@
 * Added new exception RequestCanceledException
 * Added the ability to cancel a request using CancelCompleter in AppHttp.apiHttpRequest
 * Changed file structure for http_client
+
+## 0.0.6
+
+* Methods for requests are moved from BaseCubit into two mixins, to use the old methods baseRequest and baseRequestOld, use the BaseCubitRequestMixinOld mixin
+* Added parameter authorizationToken in AppHttp.apiHttpRequest
+* Added test for BaseCubit